fix(review): keep the FIPS table out of the proto-freshness input closure
The OpenAPI example injector reads the GDELT materializer's source, so every
module the materializer imports is a `make generate` input. Importing the
FIPS table from the conflict-bulk module dragged the seed runtime (seed
utils, envelope, telemetry) into that closure and failed the proto-freshness
input contract in CI. The table now lives in the dependency-free
scripts/_gdelt-country-codes.mjs, re-exported from its old home; the new
file is registered in the pre-push proto pathspecs, the CI codegen input
list, and both Railway services' watch patterns.
