+++ scripts/railway-services.json
@@ -1191,6 +1191,7 @@
       "scripts/_conflict-hapi.mjs",
       "scripts/_country-resolver.mjs",
       "scripts/_gdelt-bulk-contract.mjs",
+      "scripts/_gdelt-country-codes.mjs",
       "scripts/_gdelt-fetch.mjs",
       "scripts/_proxy-utils.cjs",
       "scripts/_seed-contract.mjs",
@@ -1434,6 +1435,7 @@
       "scripts/_conflict-gdelt.mjs",
       "scripts/_gdelt-bulk-contract.mjs",
       "scripts/_gdelt-bulk-materializer.mjs",
+      "scripts/_gdelt-country-codes.mjs",
       "scripts/_html-entities.mjs",
       "scripts/_proxy-utils.cjs",
       "scripts/_seed-contract.mjs",
